+++ lib/repositories/src/ngxs-data-entity-collections/ngxs-data-entity-collections.repository.ts
@@ -1,24 +1,21 @@
+import {
+    EmptyDictionary,
+    EntityCollections,
+    EntityComparator,
+    EntityDictionary,
+    EntityIdType,
+    EntitySortBy,
+    EntityUpdate,
+    KeysDictionary
+} from '@angular-ru/common/entity';
 import { sortByAsc, sortByDesc } from '@angular-ru/common/object';
-import { Any } from '@angular-ru/common/typings';
+import { Any, PrimaryKey, SortOrderType } from '@angular-ru/common/typings';
 import { isNil } from '@angular-ru/common/utils';
 import { Injectable, isDevMode } from '@angular/core';
 import { Computed, DataAction, Payload } from '@ngxs-labs/data/decorators';
 import { ensureDataStateContext, ensureSnapshot } from '@ngxs-labs/data/internals';
 import { NGXS_DATA_EXCEPTIONS } from '@ngxs-labs/data/tokens';
-import {
-    EmptyDictionary,
-    EntityComparator,
-    EntityContext,
-    EntityDictionary,
-    EntityIdType,
-    EntityRepository,
-    EntitySortBy,
-    EntitySortByOrder,
-    EntityUpdate,
-    KeysDictionary,
-    NgxsEntityCollections,
-    PRIMARY_KEY
-} from '@ngxs-labs/data/typings';
+import { EntityContext, EntityRepository } from '@ngxs-labs/data/typings';
 import { ActionType, StateContext } from '@ngxs/store';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
@@ -30,13 +27,13 @@ export abstract class AbstractNgxsDataEntityCollectionsRepository<
     V,
     K extends number | string = EntityIdType,
     C = Record<string, Any>
-> extends AbstractRepository<NgxsEntityCollections<V, K, C>> implements EntityRepository<V, K, C> {
-    public primaryKey: string = PRIMARY_KEY.ID;
+> extends AbstractRepository<EntityCollections<V, K, C>> implements EntityRepository<V, K, C> {
+    public primaryKey: string = PrimaryKey.ID;
     public comparator: EntityComparator<V> | null = null;
     private readonly context!: EntityContext<V, K, C>;
 
     @Computed()
-    public get snapshot(): NgxsEntityCollections<V, K, C> {
+    public get snapshot(): EntityCollections<V, K, C> {
         return ensureSnapshot(this.getState());
     }
 
@@ -52,17 +49,17 @@ export abstract class AbstractNgxsDataEntityCollectionsRepository<
 
     @Computed()
     public get ids$(): Observable<K[]> {
-        return this.state$.pipe(map((value: NgxsEntityCollections<V, K, C>): K[] => value.ids));
+        return this.state$.pipe(map((value: EntityCollections<V, K, C>): K[] => value.ids));
     }
 
     @Computed()
     public get entities$(): Observable<EntityDictionary<K, V>> {
-        return this.state$.pipe(map((value: NgxsEntityCollections<V, K, C>): EntityDictionary<K, V> => value.entities));
+        return this.state$.pipe(map((value: EntityCollections<V, K, C>): EntityDictionary<K, V> => value.entities));
     }
 
     protected get ctx(): EntityContext<V, K, C> {
-        return ensureDataStateContext<NgxsEntityCollections<V, K, C>, StateContext<NgxsEntityCollections<V, K, C>>>(
-            this.context as StateContext<NgxsEntityCollections<V, K, C>>
+        return ensureDataStateContext<EntityCollections<V, K, C>, StateContext<EntityCollections<V, K, C>>>(
+            this.context as StateContext<EntityCollections<V, K, C>>
         );
     }
 
@@ -75,7 +72,7 @@ export abstract class AbstractNgxsDataEntityCollectionsRepository<
         return this.ctx.dispatch(actions);
     }
 
-    public getState(): NgxsEntityCollections<V, K, C> {
+    public getState(): EntityCollections<V, K, C> {
         return this.ctx.getState();
     }
 
@@ -88,7 +85,7 @@ export abstract class AbstractNgxsDataEntityCollectionsRepository<
     }
 
     public selectAll(): V[] {
-        const state: NgxsEntityCollections<V, K, C> = this.getState();
+        const state: EntityCollections<V, K, C> = this.getState();
         return state.ids.map((id: K): V => state.entities[id] as V);
     }
 
@@ -188,7 +185,7 @@ export abstract class AbstractNgxsDataEntityCollectionsRepository<
     }
 
     protected addEntityOne(entity: V): void {
-        const state: NgxsEntityCollections<V, K, C> = this.getState();
+        const state: EntityCollections<V, K, C> = this.getState();
         const id: K = this.selectIdValue(entity);
 
         if (id in state.entities) {
@@ -203,7 +200,7 @@ export abstract class AbstractNgxsDataEntityCollectionsRepository<
     }
 
     protected addEntitiesMany(entities: V[]): void {
-        const state: NgxsEntityCollections<V, K, C> = this.getState();
+        const state: EntityCollections<V, K, C> = this.getState();
         const dictionary: EntityDictionary<K, V> | EmptyDictionary<K, V> = {};
         const ids: K[] = [];
 
@@ -228,7 +225,7 @@ export abstract class AbstractNgxsDataEntityCollectionsRepository<
     }
 
     protected setEntitiesAll(entities: V[]): void {
-        const state: NgxsEntityCollections<V, K, C> = this.getState();
+        const state: EntityCollections<V, K, C> = this.getState();
         const dictionary: EntityDictionary<K, V> | EmptyDictionary<K, V> = {};
         const ids: K[] = [];
 
@@ -247,7 +244,7 @@ export abstract class AbstractNgxsDataEntityCollectionsRepository<
     }
 
     protected setEntityOne(entity: V): void {
-        const state: NgxsEntityCollections<V, K, C> = this.getState();
+        const state: EntityCollections<V, K, C> = this.getState();
         const id: K = this.selectIdValue(entity);
 
         if (id in state.entities) {
@@ -264,7 +261,7 @@ export abstract class AbstractNgxsDataEntityCollectionsRepository<
     }
 
     protected updateEntitiesMany(updates: EntityUpdate<V, K>[]): void {
-        const state: NgxsEntityCollections<V, K, C> = this.getState();
+        const state: EntityCollections<V, K, C> = this.getState();
         updates = updates.filter((update: EntityUpdate<V, K>): boolean => update.id in state.entities);
         if (updates.length === 0) {
             return;
@@ -290,7 +287,7 @@ export abstract class AbstractNgxsDataEntityCollectionsRepository<
     }
 
     protected upsertEntitiesMany(entities: V[]): void {
-        const state: NgxsEntityCollections<V, K, C> = this.getState();
+        const state: EntityCollections<V, K, C> = this.getState();
         const updates: EntityUpdate<V, K>[] = [];
         const added: V[] = [];
 
@@ -308,7 +305,7 @@ export abstract class AbstractNgxsDataEntityCollectionsRepository<
     }
 
     protected removeEntitiesMany(ids: K[]): void {
-        const state: NgxsEntityCollections<V, K, C> = this.getState();
+        const state: EntityCollections<V, K, C> = this.getState();
         const keys: KeysDictionary<K> = this.generateKeyMap(state);
         const entities: EntityDictionary<K, V> = { ...state.entities };
 
@@ -322,7 +319,7 @@ export abstract class AbstractNgxsDataEntityCollectionsRepository<
         this.setEntitiesState({ ...state, ids: state.ids.filter((id: K): boolean => id in keys), entities });
     }
 
-    protected setEntitiesState(state: NgxsEntityCollections<V, K, C>): void {
+    protected setEntitiesState(state: EntityCollections<V, K, C>): void {
         const ids: K[] = this.sortKeysByComparator(state.ids, state.entities);
         this.ctx.setState({ ...state, ids, entities: state.entities });
     }
@@ -344,11 +341,11 @@ export abstract class AbstractNgxsDataEntityCollectionsRepository<
 
     private sortByComparatorOptions(ids: K[], comparator: EntitySortBy<V>, entities: EntityDictionary<K, V>): K[] {
         switch (comparator?.sortByOrder) {
-            case EntitySortByOrder.ASC:
+            case SortOrderType.ASC:
                 return ids.sort((a: K, b: K): number =>
                     sortByAsc(comparator?.sortBy, entities[a] as V, entities[b] as V)
                 );
-            case EntitySortByOrder.DESC:
+            case SortOrderType.DESC:
                 return ids.sort((a: K, b: K): number =>
                     sortByDesc(comparator?.sortBy, entities[a] as V, entities[b] as V)
                 );
@@ -361,7 +358,7 @@ export abstract class AbstractNgxsDataEntityCollectionsRepository<
         }
     }
 
-    private generateKeyMap(state: NgxsEntityCollections<V, K, C>): KeysDictionary<K> {
+    private generateKeyMap(state: EntityCollections<V, K, C>): KeysDictionary<K> {
         return state.ids.reduce((keyDictionary: KeysDictionary<K>, id: K): KeysDictionary<K> => {
             keyDictionary[id] = id;
             return keyDictionary;